+++ src/MulleFoundation-startup.m
@@ -1,0 +1,13 @@
+//
+//  Foundation.m
+//  MulleFoundation
+//
+//  Created by Nat! on 04.04.16.
+//  Copyright © 2016 Mulle kybernetiK. All rights reserved.
+//
+#import "MulleFoundation.h"
+
+#import "MulleFoundationSetup.h"
+
+#include "MulleFoundation-startup.inc"
+