+++ RELEASENOTES.md
@@ -1,5 +1,6 @@
 ## 0.13.0
 
+* move Foundation.h to its own project
 * migrated to mulle-sde